% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshfs.R
\name{sshfs_mount}
\alias{sshfs_mount}
\alias{sshfs_unmount}
\title{Mounts and unmounts a remote folder}
\usage{
sshfs_mount(user, remote_server, remote_folder, local_folder,
  permission = c("r", "rw"))

sshfs_unmount(mount_point)
}
\arguments{
\item{user}{user}

\item{remote_server}{remote server or tunnel object returned by [tunnel_open]}

\item{remote_folder}{server's folder to be mounted}

\item{local_folder}{local folder where to mount}

\item{permission}{either "r" or "rw"}

\item{mount_point}{mount_point object returned from [sshfs_mount]}
}
\description{
Mounts and unmounts a remote folder
}
\examples{
\dontrun{
mount <- sshfs_mount(user = "myself",
                     remote_server = "server.domain",
                     remote_folder = "~/data_folder",
                     local_folder = "./DATA)

# Now you can access your files with
list.files(mount$folder)

# Unmount when done
sshfs_unmount(mount)

# If you need to tunnel over another server, first create the
# tunnel and then pass the object as a `remote_server`
tunnel <- tunnel_open(user = "myself",
                      tunnel_server = "portal.domain",
                      remote_server = "server.domain")

mount <- sshfs_mount(user = "myself",
                     remote_server = tunnel,
                     remote_folder = "~/data_folder",
                     local_folder = "./DATA)

# When done, unmount and close the tunnel.
sshfs_unmount(mount)
tunnel_close(tunnel)

}

}
