% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshfs.R
\name{tunnel_open}
\alias{tunnel_open}
\alias{tunnel_close}
\title{Opens up a ssh tunnel}
\usage{
tunnel_open(user, tunnel_server, remote_server, local_port = 4567,
  remote_port = 22)

tunnel_close(tunnel)
}
\arguments{
\item{user}{username}

\item{tunnel_server}{tunnel server}

\item{remote_server}{remote server}

\item{local_port}{local port to use for tunneling}

\item{remote_port}{remote port used for ssh}

\item{tunnel}{tunnel object returned by [tunnel_open]}
}
\description{
Opens up a ssh tunnel
}
