% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-mount.R
\name{with.sshfs_mount_point}
\alias{with.sshfs_mount_point}
\title{Evaluate expressions with mounted directory}
\usage{
\method{with}{sshfs_mount_point}(data, expr, ...)
}
\arguments{
\item{data}{a mount object returned by [sshfs_mount()]}

\item{expr}{any R expression thet will be evaluated.}

\item{...}{ignored}
}
\description{
SSHFS can be a bit finicky and fail when reading lots of files. It can be safer, then,
to mount and unmount the folder each time you read each file.
}
\details{
The function mounts the remote folder, evaluates the expresion and then unmounts.
The `expr` can access the locally-mounted folder with the `.folder` variable.
}
\examples{
\dontrun{
mount <- sshfs_mount("user", "remote.server.com", "~/data_folder", "~/loca_folder")
bici <- with(mount, read.csv(file.path(.folder, "some_file.csv")))
}
}
