% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sshfs.R
\name{sshfs_mount}
\alias{sshfs_mount}
\alias{sshfs_unmount}
\title{Mounts and unmounts a remote folder}
\usage{
sshfs_mount(
  user,
  remote_server,
  remote_folder,
  local_folder,
  permission = c("r", "rw"),
  extra_args = ""
)

sshfs_unmount(mount_point)
}
\arguments{
\item{user}{user}

\item{remote_server}{remote server or tunnel object returned by [tunnel_open]}

\item{remote_folder}{server's folder to be mounted}

\item{local_folder}{local folder where to mount}

\item{permission}{either "r" or "rw"}

\item{extra_args}{character string of extra arguments passed to `sshfs`.}

\item{mount_point}{mount_point object returned from [sshfs_mount]}
}
\description{
Mounts and unmounts a remote folder
}
